Fixed implementation of SCA assembly XML writers. Simplified most readers and writers. Added code to read/write policy intents and policy sets on relevant SCA assembly XML elements.
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/java/sca@572267 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/implementation-data/src/test/java/org/apache/tuscany/sca/implementation/data/DATAArtifactProcessorTestCase.java
@@ -133,6 +133,6 @@ public class DATAArtifactProcessorTestCase extends TestCase {
         assertNotNull(connProperties);
         assertEquals("driverClass",connProperties.getDriverClass());
         assertEquals("databaseURL",connProperties.getDatabaseURL());
-        assertEquals(1,connProperties.getLoginTimeout());
+        assertEquals(1,connProperties.getLoginTimeout().intValue());
     }    
 }
